🧹 Cleanup Exports, Change Shell to SVGIcon, Consolidate Imports
Since Shell was serving as a default wrapper for all the icons, it made sense to just rename is SVGIcon and move the type definition into the same file, then import both in each icon file.

All the icons are now exported from index.ts in a few different ways. You can import any icon individually by name, or as a group by `ColorIcons`, `DarkIcons`, `LightIcons`, or get all of them in one object with `Icons`. Webpack will properly tree-shake away any unused exports, so it's okay to use imports and define consts for grouping icons like this. Those imports should all be shaken away so long as a group isn't imported and used somewhere (in most cases we'd avoid doing this anyways).